Use trampoline when loading kernel
This avoids the issue where the currently-executing code (i.e. the jmp to the kernel) will no longer exist in the page table after setting CR3. Instead, a tiny trampoline (a `mov cr3` and a `jmp`) is allocated which is mapped in both the bootloader and the kernel's page table at the same address. The trampoline is always allocated at an address that is lower than anything used by the kernel (which is in the "high half" of the address space), so it does not collide with the virtual address space of the kernel.